Move BaseError to parser crate
