+++ core/src/lib.rs
@@ -1,9 +1,7 @@
 #![doc(html_logo_url = "https://raw.githubusercontent.com/RustPython/RustPython/main/logo.png")]
 #![doc(html_root_url = "https://docs.rs/rustpython-parser-core/")]
 
-mod error;
 mod format;
 pub mod mode;
 
-pub use error::BaseError;
 pub use mode::Mode;
